+++ include/shared_resources.h
